+++ src/proxy/src/helpers/guards.rs
@@ -88,6 +88,8 @@ pub fn is_developer() -> Result<(), String> {
         "syzio-xu6ca-burmx-4afo2-ojpcw-e75j3-m67o5-s5bes-5vvsv-du3t4-wae",
         // Olek
         "bgykr-qmmrw-bynrn-ffwva-j6th7-juxki-het4d-5sac4-7v4t2-re73t-bqe",
+        // Monitor
+        "6or45-oyaaa-aaaap-absua-cai",
     ];
 
     if developers.contains(&caller().to_text().as_str()) {