+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - "bad_request" error to "duplicate" error
+- websocket url
 
 ## [0.2.2]