+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.4]
+
+### Added
+
+- monitor principal to "is_developer" guard
+- replace space by dash when for name param in "get_group_by_name"
+
+### Changed
+
+- "IdKind::RewardBuffer" for id increment
+- get_latest_logs guard from "monitor" to "is_developer"
+
 ## [0.2.3]
 
 ### Added